Moved common mixin code to the correct directory
