+++ common/src/main/java/org/valkyrienskies/mod/mixin/util/math/MixinMatrix4f.java
@@ -1,11 +1,11 @@
-package org.valkyrienskies.mod.common.mixin.util.math;
+package org.valkyrienskies.mod.mixin.util.math;
 
 import net.minecraft.util.math.Matrix4f;
 import org.joml.Matrix4dc;
 import org.joml.Matrix4fc;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
-import org.valkyrienskies.mod.common.MixinInterfaces;
+import org.valkyrienskies.mod.MixinInterfaces;
 
 @Mixin(Matrix4f.class)
 public class MixinMatrix4f implements MixinInterfaces.ISetMatrix4fFromJOML {
